refactor: divide cmd and main
