+++ cmd/uniflow/main.go
@@ -6,6 +6,7 @@ import (
 	"os"
 	"strings"
 
+	"github.com/siyul-park/uniflow/cmd/uniflow/uniflow"
 	"github.com/siyul-park/uniflow/pkg/database"
 	"github.com/siyul-park/uniflow/pkg/database/memdb"
 	"github.com/siyul-park/uniflow/pkg/database/mongodb"
@@ -80,7 +81,7 @@ func execute() error {
 	}
 	systemx.AddToScheme(st)(sc)
 
-	cmd := NewCmd(Config{
+	cmd := uniflow.NewCmd(uniflow.Config{
 		Scheme:   sc,
 		Hook:     hk,
 		Database: db,
