+++ cmd/uniflow/uniflow/cmd.go
@@ -1,4 +1,4 @@
-package main
+package uniflow
 
 import (
 	"io/fs"
